[darwin] network: move implementation to platform folder.
