+++ cmake/treedata/ios/subdirs.txt
@@ -2,13 +2,13 @@ xbmc/platform/linux              platform/linux
 xbmc/platform/linux/network      platform/linux/network
 xbmc/input/touch                 input/touch
 xbmc/input/touch/generic         input/touch/generic
-xbmc/network/osx                 network/osx
 xbmc/peripherals/bus/osx         peripherals/bus/osx
 xbmc/powermanagement/osx         powermanagement/osx
 xbmc/platform/posix              posix
 xbmc/platform/darwin             platform/darwin
 xbmc/platform/darwin/ios         platform/ios
 xbmc/platform/darwin/ios-common  platform/ios-common
+xbmc/platform/darwin/osx/network  platform/osx/network
 xbmc/platform/darwin/osx/storage  platform/osx/storage
 xbmc/filesystem/posix            filesystem/posix
 xbmc/utils/posix                 utils/posix
