+++ cmake/treedata/osx/subdirs.txt
@@ -1,11 +1,11 @@
 xbmc/platform/linux          platform/linux
 xbmc/platform/linux/network  platform/linux/network
-xbmc/network/osx             network/osx
 xbmc/peripherals/bus/osx     peripherals/bus/osx
 xbmc/powermanagement/osx     powermanagement/osx
 xbmc/platform/posix          posix
 xbmc/platform/darwin         platform/darwin
 xbmc/platform/darwin/osx     platform/osx
+xbmc/platform/darwin/osx/network  platform/osx/network
 xbmc/platform/darwin/osx/storage  platform/osx/storage
 xbmc/filesystem/posix        filesystem/posix
 xbmc/utils/posix             utils/posix
