+++ xbmc/platform/darwin/osx/network/CMakeLists.txt
@@ -5,4 +5,4 @@ set(HEADERS ioshacks.h
             ZeroconfBrowserOSX.h
             ZeroconfOSX.h)
 
-core_add_library(network_osx)
+core_add_library(platform_osx_network)
