+++ xbmc/platform/darwin/osx/network/ZeroconfBrowserOSX.h
